requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
coverage==4.5.1
Django==2.0.2
django-betterforms==1.1.4
django-sslserver==0.20
idna==2.6
kombu==4.1.0
PyJWT==1.6.0
PySocks==1.6.8
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
twilio==6.10.4
urllib3==1.22
vine==1.1.4
